requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
coverage==7.6.12
decorator==5.2.1
distro==1.9.0
Django==5.1.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
executing==2.2.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipython==8.34.0
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
